refactor: consolidate all exports into a single file (index.js) inside config folder
