+++ src/components/faq/faq.styles.jsx
@@ -5,8 +5,8 @@ export const FaqContainer = styled.div`
   ${tw`bg-brown-2 mb-20 3md:mb-32 pl-3`}
 `
 export const FaqHeading = styled.div`
-  ${tw`font-Poppins text-brown-4 text-4.5xl font-bold 3md:text-9xl 2.5md:text-7.5xl 0.7xsm:text-6xl`}
+  ${tw`font-Poppins text-brown-4 text-4.5xl font-bold 3md:text-8.5xl 2.5md:text-7.5xl 0.7xsm:text-6xl`}
 `
 export const FaqContent = styled.div`
-  ${tw`font-Poppins text-brown-4 text-xl font-bold 3md:text-6xl 2.5md:text-5.5xl 0.7xsm:text-4.5xl flex justify-center mt-10`}
+  ${tw`font-Poppins text-brown-4 text-xl font-bold 3md:text-5.5xl 2.5md:text-5.5xl 0.7xsm:text-4.5xl flex justify-center mt-10`}
 `
