+++ src/components/marginals/Nav_Hero/Description.jsx
@@ -6,7 +6,7 @@ import {
   DescriptionParagraph,
   DescriptionTitle
 } from './Description.styles'
-import { DescriptionContent } from '../../../config/content/navData/Nav_Hero'
+import { DescriptionContent } from '../../../config/index'
 import { ButtonWrapper } from './Description.styles'
 
 const Description = () => {
