+++ src/components/marginals/Nav_Hero/Nav.styles.jsx
@@ -1,6 +1,6 @@
 import tw from 'twin.macro'
 import styled from 'styled-components'
-import { links } from '../../../config/content/navData/Nav_Hero'
+import { links } from '../../../config/index'
 
 const NavCont = styled.nav`
   ${tw`
