+++ src/components/registrationHelp/registrationhelp.styles.jsx
@@ -1,6 +1,6 @@
 import tw from 'twin.macro'
 import styled from 'styled-components'
-import { apostrophe } from '../../config/content/registrationHelp/registrationHelp'
+import { apostrophe } from '../../config/index'
 
 export const RegistrationHelpContainer = styled.div`
   ${tw`bg-brown-2 mb-14 0.7xsm:mb-14 flex flex-col justify-center items-center`}
