+++ src/components/schedule/schedule.styles.jsx
@@ -2,11 +2,11 @@ import tw from 'twin.macro'
 import styled from 'styled-components'
 
 export const Container = styled.div`
-  ${tw`bg-brown-2 mb-20 3md:mb-32 pl-3`}
+  ${tw`bg-brown-2 mb-12 3md:mb-20 pl-3 sm:pl-6`}
 `
 export const Heading = styled.div`
-  ${tw`font-Poppins text-brown-4 text-4.5xl font-bold 3md:text-9xl 2.5md:text-7.5xl 0.7xsm:text-6xl`}
+  ${tw`font-Poppins text-brown-4 text-4.5xl font-bold 3md:text-8.5xl 2.5md:text-7.5xl 0.7xsm:text-6xl`}
 `
 export const Content = styled.div`
-  ${tw`font-Poppins text-brown-4 text-xl font-bold 3md:text-6xl 2.5md:text-5.5xl 0.7xsm:text-4.5xl flex justify-center mt-10`}
+  ${tw`font-Poppins text-brown-4 text-xl font-bold 3md:text-5.5xl 2.5md:text-5.5xl 0.7xsm:text-4.5xl flex justify-center mt-9`}
 `
