+++ src/components/teamMembers/teamSection.styles.jsx
@@ -2,7 +2,7 @@ import tw from 'twin.macro'
 import styled from 'styled-components'
 
 export const TeamSectionContainer = styled.div`
-  ${tw`bg-brown-2 pl-3 pr-3 mb-24 mt-20`}
+  ${tw`bg-brown-2 pl-3 pr-3 mb-14 mt-20`}
 `
 
 export const Loading = styled.h2`
