+++ src/config/index.js
@@ -1,0 +1,27 @@
+// Desc: Content data for the website
+
+// Team members data
+export { default as TeamMembers } from './content/teamsData/team'
+
+// Sponsors data
+export { default as SponsorsData } from './content/sponsorsData/sponsors'
+
+// Registraion help data
+import { apostrophe } from './content/registrationHelp/registrationHelp'
+export { apostrophe }
+
+// Leads data
+import { Lead, SRCchairs } from './content/leads/leads'
+export { Lead, SRCchairs }
+
+// Nav data
+import { links, navLinks, DescriptionContent, heroContent } from './content/navData/Nav_Hero'
+export { links, navLinks, DescriptionContent, heroContent }
+
+// Events data
+import { events, seeMoreIcon, XIcon, nextArrowIcon, prevArrowIcon } from './content/events/events'
+export { events, seeMoreIcon, XIcon, nextArrowIcon, prevArrowIcon }
+
+// AboutUs data
+import { data, srcData, data2, aicheData, aboutImage } from './content/about/about'
+export { data, srcData, data2, aicheData, aboutImage }
